Add waitForServiceNotUnavailable method to TestUtils class
A new method, named waitForServiceNotUnavailable, was introduced in TestUtils class. This method uses an HttpClient to poll a given HTTP URL until it receives a response other than a 503 (Service Unavailable). This method was also used in the ModelServingST and DistributedST classes, replacing previously hard-coded segments of code that performed the same action.